+++ lib/RgsCertificate.php
@@ -69,4 +69,9 @@ class RgsCertificate
 
         return $result;
     }
+
+    public function isSslClient(string $certificate): bool
+    {
+        return openssl_x509_checkpurpose($certificate, X509_PURPOSE_SSL_CLIENT, [$this->validca_path]);
+    }
 }