+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Permettre de poster les Pes Acquit Retour #1156
 - Permettre de configurer la configuration Apache SSLInsecureRenegotiation #1244
+- Afficher un message d'erreur lorsqu'un certificat utilisateur ne semble pas permettre l'authentification #1023
 
 ### Corrections